change the message in controller
