+++ Go-Project/router.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/RaymondCode/simple-demo/controller"
 	"github.com/gin-gonic/gin"
+	"github.com/life-studied/douyin-simple/controller"
 )
 
 func initRouter(r *gin.Engine) {
