+++ Go-Project/service/message.go
@@ -3,11 +3,27 @@ package service
 import (
 	"encoding/json"
 	"fmt"
-	"github.com/RaymondCode/simple-demo/controller"
 	"io"
 	"net"
 	"sync"
 )
+
+type Message struct {
+	Id         int64  `json:"id,omitempty"`
+	Content    string `json:"content,omitempty"`
+	CreateTime string `json:"create_time,omitempty"`
+}
+
+type MessageSendEvent struct {
+	UserId     int64  `json:"user_id,omitempty"`
+	ToUserId   int64  `json:"to_user_id,omitempty"`
+	MsgContent string `json:"msg_content,omitempty"`
+}
+
+type MessagePushEvent struct {
+	FromUserId int64  `json:"user_id,omitempty"`
+	MsgContent string `json:"msg_content,omitempty"`
+}
 
 var chatConnMap = sync.Map{}
 
@@ -43,7 +59,7 @@ func process(conn net.Conn) {
 			continue
 		}
 
-		var event = controller.MessageSendEvent{}
+		var event = MessageSendEvent{}
 		_ = json.Unmarshal(buf[:n], &event)
 		fmt.Printf("Receive Message：%+v\n", event)
 
@@ -60,7 +76,7 @@ func process(conn net.Conn) {
 			continue
 		}
 
-		pushEvent := controller.MessagePushEvent{
+		pushEvent := MessagePushEvent{
 			FromUserId: event.UserId,
 			MsgContent: event.MsgContent,
 		}
